Push docker images built in PRs to the registry as well.
It can be useful during reviews to access the Docker image that is built
in a PR/branch after a change to the toolchain. Currently those images
were not pushed but only built during the action and then discarded.

Now we push them to the registery using a different tag, and we use
the correct tag while building libdragon as well (either the tag
of the current branch/PR, or "latest" in case no image was built).